+++ contracts/ownership/Shareable.sol
@@ -11,18 +11,18 @@ contract Shareable {
 
   // struct for the status of a pending operation.
   struct PendingState {
-    uint yetNeeded;
-    uint ownersDone;
-    uint index;
+    uint256 yetNeeded;
+    uint256 ownersDone;
+    uint256 index;
   }
 
   // the number of owners that must confirm the same operation before it is run.
-  uint public required;
+  uint256 public required;
 
   // list of owners
   address[256] owners;
   // index on the list of owners to allow reverse lookup
-  mapping(address => uint) ownerIndex;
+  mapping(address => uint256) ownerIndex;
   // the ongoing operations.
   mapping(bytes32 => PendingState) pendings;
   bytes32[] pendingsIndex;
@@ -59,10 +59,10 @@ contract Shareable {
    * @param _owners A list of owners.
    * @param _required The amount required for a transaction to be approved.
    */
-  function Shareable(address[] _owners, uint _required) {
+  function Shareable(address[] _owners, uint256 _required) {
     owners[1] = msg.sender;
     ownerIndex[msg.sender] = 1;
-    for (uint i = 0; i < _owners.length; ++i) {
+    for (uint256 i = 0; i < _owners.length; ++i) {
       owners[2 + i] = _owners[i];
       ownerIndex[_owners[i]] = 2 + i;
     }
@@ -77,12 +77,12 @@ contract Shareable {
    * @param _operation A string identifying the operation.
    */
   function revoke(bytes32 _operation) external {
-    uint index = ownerIndex[msg.sender];
+    uint256 index = ownerIndex[msg.sender];
     // make sure they're an owner
     if (index == 0) {
       return;
     }
-    uint ownerIndexBit = 2**index;
+    uint256 ownerIndexBit = 2**index;
     var pending = pendings[_operation];
     if (pending.ownersDone & ownerIndexBit > 0) {
       pending.yetNeeded++;
@@ -93,10 +93,10 @@ contract Shareable {
 
   /**
    * @dev Gets an owner by 0-indexed position (using numOwners as the count)
-   * @param ownerIndex Uint The index of the owner
+   * @param ownerIndex uint256 The index of the owner
    * @return The address of the owner
    */
-  function getOwner(uint ownerIndex) external constant returns (address) {
+  function getOwner(uint256 ownerIndex) external constant returns (address) {
     return address(owners[ownerIndex + 1]);
   }
 
@@ -117,7 +117,7 @@ contract Shareable {
    */
   function hasConfirmed(bytes32 _operation, address _owner) constant returns (bool) {
     var pending = pendings[_operation];
-    uint index = ownerIndex[_owner];
+    uint256 index = ownerIndex[_owner];
 
     // make sure they're an owner
     if (index == 0) {
@@ -125,7 +125,7 @@ contract Shareable {
     }
 
     // determine the bit to set for this owner.
-    uint ownerIndexBit = 2**index;
+    uint256 ownerIndexBit = 2**index;
     return !(pending.ownersDone & ownerIndexBit == 0);
   }
 
@@ -136,7 +136,7 @@ contract Shareable {
    */
   function confirmAndCheck(bytes32 _operation) internal returns (bool) {
     // determine what index the present sender is:
-    uint index = ownerIndex[msg.sender];
+    uint256 index = ownerIndex[msg.sender];
     // make sure they're an owner
     if (index == 0) {
       throw;
@@ -153,7 +153,7 @@ contract Shareable {
       pendingsIndex[pending.index] = _operation;
     }
     // determine the bit to set for this owner.
-    uint ownerIndexBit = 2**index;
+    uint256 ownerIndexBit = 2**index;
     // make sure we (the message sender) haven't confirmed this operation previously.
     if (pending.ownersDone & ownerIndexBit == 0) {
       Confirmation(msg.sender, _operation);
@@ -177,8 +177,8 @@ contract Shareable {
    * @dev Clear the pending list.
    */
   function clearPending() internal {
-    uint length = pendingsIndex.length;
-    for (uint i = 0; i < length; ++i) {
+    uint256 length = pendingsIndex.length;
+    for (uint256 i = 0; i < length; ++i) {
       if (pendingsIndex[i] != 0) {
         delete pendings[pendingsIndex[i]];
       }
