+++ contracts/payment/PullPayment.sol
@@ -10,17 +10,17 @@ import '../SafeMath.sol';
  * contract and use asyncSend instead of send.
  */
 contract PullPayment {
-  using SafeMath for uint;
+  using SafeMath for uint256;
 
-  mapping(address => uint) public payments;
-  uint public totalPayments;
+  mapping(address => uint256) public payments;
+  uint256 public totalPayments;
 
   /**
   * @dev Called by the payer to store the sent amount as credit to be pulled.
   * @param dest The destination address of the funds.
   * @param amount The amount to transfer.
   */
-  function asyncSend(address dest, uint amount) internal {
+  function asyncSend(address dest, uint256 amount) internal {
     payments[dest] = payments[dest].add(amount);
     totalPayments = totalPayments.add(amount);
   }
@@ -30,7 +30,7 @@ contract PullPayment {
   */
   function withdrawPayments() {
     address payee = msg.sender;
-    uint payment = payments[payee];
+    uint256 payment = payments[payee];
 
     if (payment == 0) {
       throw;
