+++ contracts/token/ERC20Basic.sol
@@ -7,8 +7,8 @@ pragma solidity ^0.4.11;
  * @dev see https://github.com/ethereum/EIPs/issues/20
  */
 contract ERC20Basic {
-  uint public totalSupply;
-  function balanceOf(address who) constant returns (uint);
-  function transfer(address to, uint value);
-  event Transfer(address indexed from, address indexed to, uint value);
+  uint256 public totalSupply;
+  function balanceOf(address who) constant returns (uint256);
+  function transfer(address to, uint256 value);
+  event Transfer(address indexed from, address indexed to, uint256 value);
 }
